Wait until the Keyguard is drawn to dismiss the boot dialog.
In a631d49, we check that the keyguard is ready to avoid displaying the
homescreen before the keyguard if the boot message is dismissed too
soon. However, this only checked mKeyguardDrawnOnce, which is set to
true before the keyguard is fully visible to the user. Instead, wait for
mKeyguardDrawComplete, which will be triggered when the Keyguard is
actually fully visible.

With this change, the keyguard should always be the first thing displayed
after the boot dialog is dismissed, not the home screen.

Change-Id: I92f7df66e04613e9f9b39d293b190fc2473de176
Signed-off-by: bgos10 <bgos10@yahoo.co.in>