Bluetooth: Update the A2DP status properly
If the device is removed from the hash table before updating
the status because of asynchronous operations involved, audio
is not able to update policy manager regarding A2dp disconnection.
This change takes care of forcibly updating HAL in such scenario
so that Audio chooses the proper device on next playback.

CRs-Fixed: 1031541
Change-Id: I4d0e56590b5e01cda6657f61d31d0b895fce0306
Signed-off-by: bgos10 <bishalgoswami@gmail.com>