MSIM: Fix to show PUK view when two sims locked on PIN.
Currently when both sims are locked on PIN, upon exhausting
PIN attempts for SIM1 PIN, SIM1 PUK view is not displayed
and the user is not able to unlock the PUK.

This is due to while fetching the getSecurityMode in Key
-GuardSecurityModel, the ordering of checking the security
mode is PIN first and PUK later.So when one sim is locked
on PUK and the other is locked on PIN the ordering of checking
the PIN first is causing to return PIN as security mode in place
of PUK. This will lead to PUK lock screen not been shown to user.

Fix: In KeyGuardSecurityModel.java, while querying the security Mode
modify the order to check for security mode i.e PUK locked state first
and then PIN locked state.

In KeyGuardSimPinView and KeyGuardSimPukView, check for sim locked on
PIN and reset the flags to show the default message.

Change-Id: I97fb81c23a94c18a37c251337ef48f62d26240a7
CRs-Fixed: 1041754
Signed-off-by: bgos10 <bishalgoswami@gmail.com>