input: change config_doubleTapOnAppSwitchKeyBehavior to KEY_ACTION_NOTHING
From Android N, Google implemented this feature themselves
and it works in a different way than a traditional double-tap gesture.

SystemUI immediately enters the recents app switcher screen
when the first recents key action is dispatched.
Then the second recents key action actually makes the SystemUI
resume to the last app, which makes it look like "Launch last app".

This has the obvious advantage of entering the recents app switcher screen
earlier since it does not have to wait for the second recents key action.

In a perfect world, we would want to remove "No action" from the Settings
and remap "Launch last app" to "No action" for recents key,
as it's the AOSP's default.

Until we figure out a clean way to do that,
just change the default to KEY_ACTION_NOTHING for recents key.

Change-Id: Ib52001d6658d8f8d3bf9efbb6f4ffbfd6f05d24e
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>

Conflicts:
	core/res/res/values/pa_config.xml